Avoid compiler warnings for signed vs unsigned integer comparisons

+++ scalar_highway_tree_hash.cc
@@ -113,7 +113,7 @@ uint64_t ScalarHighwayTreeHash(const Lanes& key, const uint8_t* bytes,
   const size_t remainder = size & (kPacketSize - 1);
   const size_t truncated_size = size - remainder;
   const uint64_t* packets = reinterpret_cast<const uint64_t*>(bytes);
-  for (int i = 0; i < truncated_size / sizeof(uint64_t); i += kNumLanes) {
+  for (size_t i = 0; i < truncated_size / sizeof(uint64_t); i += kNumLanes) {
     state.Update(packets + i);
   }
 
@@ -121,7 +121,7 @@ uint64_t ScalarHighwayTreeHash(const Lanes& key, const uint8_t* bytes,
   const size_t remainder_mod4 = remainder & 3;
   uint32_t packet4 = static_cast<uint32_t>(size) << 24;
   const uint8_t* final_bytes = bytes + size - remainder_mod4;
-  for (int i = 0; i < remainder_mod4; ++i) {
+  for (size_t i = 0; i < remainder_mod4; ++i) {
     packet4 += static_cast<uint32_t>(final_bytes[i]) << (i * 8);
   }
